Added new Feature (Permalink & Make your own card) User obliged to enter their name each time they wanted to see the card, I turned the website into a "card generator" that will make you generate your name once and get a link to it to always display the name's card associated with the url...
I also added a "Make your own!" button at the bottom of the card for interested people who enjoyed the card message and wanted to use the website to make their own card too.

Added an "Enter" key listener to the name input to generate the card additionally with just clicking the button at the form.